+++ ndhistogram/src/value/mean.rs
@@ -5,8 +5,6 @@ use std::{
 
 use num_traits::{abs, Float, NumOps, One, Signed};
 
-use serde::{Deserialize, Serialize};
-
 use crate::FillWith;
 
 /// This ndhistogram bin value computes the mean of the data samples provided when
@@ -36,9 +34,8 @@ use crate::FillWith;
 /// let mean = hist.value(&0.0);
 /// assert_eq!(mean.unwrap().get(), 2.0); // should be the mean of [1,2,3]
 /// ```
-#[derive(
-    Copy, Default, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Serialize, Deserialize,
-)]
+#[derive(Copy, Default, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Mean<T = f64, O = f64, C = u32> {
     sumw: T,
     sumw2: T,