+++ ndhistogram/src/value/sum.rs
@@ -1,14 +1,13 @@
 use num_traits::Float;
 
 use crate::Fill;
-use serde::{Deserialize, Serialize};
+use crate::FillWith;
 
 /// ndhistogram bin value type for filling unweighted values.
 /// Analogous to [WeightedSum](crate::value::WeightedSum). Methods returning variance and standard
 /// deviation assume Poisson statistics.
-#[derive(
-    Copy, Default, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Serialize, Deserialize,
-)]
+#[derive(Copy, Default, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Sum<T = f64> {
     sum: T,
 }
@@ -51,5 +50,16 @@ impl<T: Copy + Fill> Fill for Sum<T> {
     #[inline]
     fn fill(&mut self) {
         self.sum.fill();
+    }
+}
+
+impl<T, W> FillWith<W> for Sum<T>
+where
+    T: FillWith<W> + Copy,
+    W: Copy,
+{
+    #[inline]
+    fn fill_with(&mut self, weight: W) {
+        self.sum.fill_with(weight);
     }
 }