+++ ndhistogram/tests/mod.rs
@@ -12,7 +12,12 @@ mod test_ndhistogram_binary_ops;
 mod test_ndhistogram_display;
 mod test_ndhistogram_macro;
 mod test_noflow_axis;
+
+#[cfg(feature = "rayon")]
+mod test_parallel_iterators;
+#[cfg(feature = "serde")]
 mod test_serialization;
+
 mod test_trait_object_safety;
 mod test_uniform_axis;
 mod test_uniform_axis_integer;